[XBOX] merged: Linuxport revisions 25036,25037,25043,25046,25059,25060,25068,25080,25091,25176,25181,25191,25192 25205,25206,25248-25255,25264,25267,25292,25304,25305,25336,25359
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/xbox@25365 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/mplayer/mplayer.cpp
@@ -1939,7 +1939,7 @@ void CMPlayer::SeekTime(__int64 iTime)
   }
   g_infoManager.m_performingSeek = false;
   WaitOnCommand();
-  m_callback.OnPlayBackSeek(iTime);
+  m_callback.OnPlayBackSeek((int)iTime);
 }
 
 //Time in milliseconds
